pierwszy commit po wielu zmianach tylko lokalnie

+++ src/main/java/pl/polskaamazonka/backend/controller/CategoryController.java
@@ -1,7 +1,10 @@
 package pl.polskaamazonka.backend.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import pl.polskaamazonka.backend.dto.CategoryDTO;
 import pl.polskaamazonka.backend.service.CategoryService;
 
@@ -20,7 +23,7 @@ public class CategoryController {
     }
 
     @GetMapping("/{id}")
-    public CategoryDTO getById(@PathVariable Integer id) {
+    public CategoryDTO getById(@PathVariable Long id) {
         return categoryService.getById(id);
     }
 }
