+++ src/main/java/pl/polskaamazonka/backend/model/Category.java
@@ -4,7 +4,10 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 
 @Getter
@@ -17,12 +20,11 @@ public class Category {
     @Id
     @ColumnDefault("nextval('category_id_seq'::regclass)")
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name", nullable = false)
     private String name;
 
     @Column(name = "image_url", length = 500)
     private String imageUrl;
-
 }
